+++ src/components/Address.tsx
@@ -1,9 +1,9 @@
 import type { StripeAddressElementChangeEvent, StripeAddressElementOptions } from '@stripe/stripe-js'
 import type { Component } from 'solid-js'
 import { mergeProps, splitProps } from 'solid-js'
-import type { ElementProps } from './types'
-import { createWrapper } from './primitives/createWrapper'
-import { createStripeElement } from './primitives/createStripeElement'
+import type { ElementProps } from '../types'
+import { createWrapper } from '../primitives/createWrapper'
+import { createStripeElement } from '../primitives/createStripeElement'
 
 export type AddressProps = ElementProps<'address', StripeAddressElementChangeEvent & { error: any }> & StripeAddressElementOptions
