+++ src/components/Ideal.tsx
