+++ src/index.tsx
@@ -1,16 +1,16 @@
-export { PaymentElement } from './payment-element/PaymentElement'
-export { Card } from './credit-cards/Card'
-export { CardNumber } from './credit-cards/CardNumber'
-export { CardCvc } from './credit-cards/CardCvc'
-export { CardExpiry } from './credit-cards/CardExpiry'
-export { PaymentRequestButton } from './googlepay-applepay/PaymentRequestButton'
-export { LinkAuthenticationElement } from './link-authentication/LinkAuthenticationElement'
-export { Ideal } from './ideal/Ideal'
-export { Iban } from './sepa/Iban'
-export { Address } from './Address'
+export { PaymentElement } from './components/PaymentElement'
+export { Card } from './components/Card'
+export { CardNumber } from './components/CardNumber'
+export { CardCvc } from './components/CardCvc'
+export { CardExpiry } from './components/CardExpiry'
+export { PaymentRequestButton } from './components/PaymentRequestButton'
+export { LinkAuthenticationElement } from './components/LinkAuthenticationElement'
+export { Ideal } from './components/Ideal'
+export { Iban } from './components/Iban'
+export { Address } from './components/Address'
 export {
   Elements,
   useStripe,
   useStripeElements,
   useStripeProxy,
-} from './Elements'
+} from './components/Elements'
